후기 연결

+++ IRecipe/app/src/main/java/com/umcproject/irecipe/domain/model/PostDetail.kt
@@ -10,7 +10,7 @@ data class PostDetail(
     val writerNick: String?,
     val writerProfileUrl: String?,
     val likes: Int?,
-    val score: Int?,
+    val score: Double?,
     val reviewCount: Int?,
     val createdAt: String?,
     val isLike: Boolean?
